+++ stock/src/main/java/com/sparta/rooibos/stock/application/aop/RoleCheck.java
@@ -1,4 +1,4 @@
-package com.sparta.rooibos.stock.application.annotation;
+package com.sparta.rooibos.stock.application.aop;
 
 import com.sparta.rooibos.stock.application.type.Role;
